training/requirements.txt
# Training Module Dependencies
# Install with: pip install -r requirements.txt

# Core ML libraries
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0

# Supabase client (for data export)
httpx>=0.25.0
python-dotenv>=1.0.0
